Update ldval size to 64 bits

+++ src/cpu/o3/cvu.hh
@@ -25,7 +25,7 @@ class CVU
         ThreadID tid;
 
         /* constant data */
-        uint8_t data;
+        uint64_t data;
 
         /** Whether or not the entry is valid. */
         bool valid;
@@ -58,14 +58,14 @@ class CVU
      *  @param data The value the load instruction should load
      *  @param tid The thread id.
      */
-    void update(Addr instPc, Addr data_addr, uint8_t data, ThreadID tid);
+    void update(Addr instPc, Addr data_addr, uint64_t data, ThreadID tid);
 
     // invalidate the corresponding entry, call upon a store instruction
     // return true if a matching entry was found
     bool invalidate(Addr instPC, Addr LwdataAddr, ThreadID tid);
 
     // replace the LRU entry, should be called if the table is full
-    void replacement(unsigned instr_idx, Addr data_addr, uint8_t data, ThreadID tid);
+    void replacement(unsigned instr_idx, Addr data_addr, uint64_t data, ThreadID tid);
 
     // keep track of the LRU, should be called everytime
     // an entry is referenced
